+++ test/setup.js
@@ -1,0 +1,2 @@
+const noop = () => undefined
+require.extensions['.css'] = noop